Compile middlewares in a single file. Fix int width for rps param
